Refactor: playlistDetail 관련 컴포넌트를 components/playlistDetail로 이동
